refactor(components): [button]
* Extract logic code out of component.
* Reorganize code for better readabilities.
