+++ packages/components/button/src/button.vue
@@ -41,17 +41,9 @@
 </template>
 
 <script lang="ts" setup>
-import { Text, computed, inject, ref, useSlots } from 'vue'
 import { ElIcon } from '@element-plus/components/icon'
-import {
-  useDeprecated,
-  useDisabled,
-  useFormItem,
-  useGlobalConfig,
-  useNamespace,
-  useSize,
-} from '@element-plus/hooks'
-import { buttonGroupContextKey } from '@element-plus/tokens'
+import { useNamespace } from '@element-plus/hooks'
+import { useButton } from './use-button'
 import { buttonEmits, buttonProps } from './button'
 import { useButtonCustomStyle } from './button-custom'
 
@@ -61,53 +53,11 @@ defineOptions({
 
 const props = defineProps(buttonProps)
 const emit = defineEmits(buttonEmits)
-const slots = useSlots()
-
-useDeprecated(
-  {
-    from: 'type.text',
-    replacement: 'link',
-    version: '3.0.0',
-    scope: 'props',
-    ref: 'https://element-plus.org/en-US/component/button.html#button-attributes',
-  },
-  computed(() => props.type === 'text')
-)
-
-const buttonGroupContext = inject(buttonGroupContextKey, undefined)
-const globalConfig = useGlobalConfig('button')
-const ns = useNamespace('button')
-const { form } = useFormItem()
-const _size = useSize(computed(() => buttonGroupContext?.size))
-const _disabled = useDisabled()
-const _ref = ref<HTMLButtonElement>()
-
-const _type = computed(() => props.type || buttonGroupContext?.type || '')
-const autoInsertSpace = computed(
-  () => props.autoInsertSpace ?? globalConfig.value?.autoInsertSpace ?? false
-)
-
-// add space between two characters in Chinese
-const shouldAddSpace = computed(() => {
-  const defaultSlot = slots.default?.()
-  if (autoInsertSpace.value && defaultSlot?.length === 1) {
-    const slot = defaultSlot[0]
-    if (slot?.type === Text) {
-      const text = slot.children as string
-      return /^\p{Unified_Ideograph}{2}$/u.test(text.trim())
-    }
-  }
-  return false
-})
 
 const buttonStyle = useButtonCustomStyle(props)
-
-const handleClick = (evt: MouseEvent) => {
-  if (props.nativeType === 'reset') {
-    form?.resetFields()
-  }
-  emit('click', evt)
-}
+const ns = useNamespace('button')
+const { _ref, _size, _type, _disabled, shouldAddSpace, handleClick } =
+  useButton(props, emit)
 
 defineExpose({
   /** @description button html element */
