+++ packages/components/button/src/instance.ts
@@ -1,0 +1,5 @@
+import type Button from './button.vue'
+import type ButtonGroup from './button-group.vue'
+
+export type ButtonInstance = InstanceType<typeof Button>
+export type ButtonGroupInstance = InstanceType<typeof ButtonGroup>
